retrait de l'interface du controller (inutile)
